ISSUE #2754 federations folder removed from processors bad merge fixes
